Extract ENet logic to separate connection files
